+++ packages/mobile/src/components/share-drawer/ShareDrawer.tsx
@@ -1,28 +1,16 @@
-import React, { useCallback, useContext, useRef, useState } from 'react'
+import React, { useCallback, useContext, useRef } from 'react'
 
-import EventEmitter from 'events'
-import path from 'path'
-
-import type { Color } from '@audius/common'
 import {
-  encodeHashId,
-  FeatureFlags,
   accountSelectors,
   collectionsSocialActions,
-  tracksSocialActions,
-  usersSocialActions,
+  FeatureFlags,
   shareModalUISelectors,
   shareSoundToTiktokModalActions,
-  uuid,
-  ErrorLevel
+  tracksSocialActions,
+  usersSocialActions
 } from '@audius/common'
 import Clipboard from '@react-native-clipboard/clipboard'
-import type { FFmpegSession } from 'ffmpeg-kit-react-native'
-import { FFmpegKit, ReturnCode } from 'ffmpeg-kit-react-native'
 import { Linking, View } from 'react-native'
-import Config from 'react-native-config'
-import RNFS from 'react-native-fs'
-import Share from 'react-native-share'
 import ViewShot from 'react-native-view-shot'
 import { useDispatch, useSelector } from 'react-redux'
 
@@ -34,20 +22,17 @@ import IconTikTokInverted from 'app/assets/images/iconTikTokInverted.svg'
 import IconTwitterBird from 'app/assets/images/iconTwitterBird.svg'
 import DeprecatedText from 'app/components/text'
 import { useFeatureFlag } from 'app/hooks/useRemoteConfig'
-import { apiClient } from 'app/services/audius-api-client'
-import { getDominantColors } from 'app/services/threads/getDominantColors'
 import { makeStyles } from 'app/styles'
-import { convertRGBToHex } from 'app/utils/convertRGBtoHex'
-import { reportToSentry } from 'app/utils/reportToSentry'
 import { Theme, useThemeColors, useThemeVariant } from 'app/utils/theme'
 
 import ActionDrawer from '../action-drawer'
-import { useTrackImage } from '../image/TrackImage'
 import { ToastContext } from '../toast/ToastContext'
 
 import { ShareToStorySticker } from './ShareToStorySticker'
 import { messages } from './messages'
+import { useShareToStory } from './useShareToStory'
 import { getContentUrl, getTwitterShareUrl } from './utils'
+
 const { getShareContent, getShareSource } = shareModalUISelectors
 const { requestOpen: requestOpenTikTokModal } = shareSoundToTiktokModalActions
 const { shareUser } = usersSocialActions
@@ -99,16 +84,9 @@ const useStyles = makeStyles(({ palette }) => ({
   }
 }))
 
-const stickerLoadedEventEmitter = new EventEmitter()
-const STICKER_LOADED_EVENT = 'loaded' as const
-
-const DEFAULT_DOMINANT_COLORS = ['000000', '434343']
-
 export const ShareDrawer = () => {
   const styles = useStyles()
   const viewShotRef = useRef() as React.RefObject<ViewShot>
-  const [shouldRenderShareToStorySticker, setShouldRenderShareToStorySticker] =
-    useState(false)
 
   const { isEnabled: isShareToTikTokEnabled } = useFeatureFlag(
     FeatureFlags.SHARE_SOUND_TO_TIKTOK
@@ -116,12 +94,6 @@ export const ShareDrawer = () => {
   const { isEnabled: isShareToInstagramStoryEnabled } = useFeatureFlag(
     FeatureFlags.SHARE_TO_STORY
   )
-
-  const isStickerImageLoadedRef = useRef(false)
-  const handleShareToStoryStickerLoad = () => {
-    isStickerImageLoadedRef.current = true
-    stickerLoadedEventEmitter.emit(STICKER_LOADED_EVENT)
-  }
 
   const { primary, secondary, neutral, staticTwitterBlue } = useThemeColors()
   const themeVariant = useThemeVariant()
@@ -136,32 +108,6 @@ export const ShareDrawer = () => {
     account &&
     account.user_id === content.artist.user_id
   const shareType = content?.type ?? 'track'
-  const { source: trackImageSource } = useTrackImage(content?.track)
-  const trackImageUri = trackImageSource[2]?.uri
-  const captureStickerImage = useCallback(async () => {
-    if (!isStickerImageLoadedRef.current) {
-      // Wait for the sticker component and image inside it to load. If this hasn't happened in 5 seconds, assume that it failed.
-      await Promise.race([
-        new Promise((resolve) =>
-          stickerLoadedEventEmitter.once(STICKER_LOADED_EVENT, resolve)
-        ),
-        new Promise((resolve) => {
-          setTimeout(resolve, 5000)
-        })
-      ])
-
-      if (!isStickerImageLoadedRef.current) {
-        // Loading the sticker failed; return undefined
-        throw new Error('The sticker component did not load successfully.')
-      }
-    }
-
-    let res: string | undefined
-    if (viewShotRef && viewShotRef.current && viewShotRef.current.capture) {
-      res = await viewShotRef.current.capture()
-    }
-    return res
-  }, [])
 
   const handleShareToTwitter = useCallback(async () => {
     if (!content) return
@@ -180,123 +126,11 @@ export const ShareDrawer = () => {
     }
   }, [content, dispatch])
 
-  const handleShareToInstagramStory = useCallback(async () => {
-    if (content?.type === 'track') {
-      if (!shouldRenderShareToStorySticker) {
-        setShouldRenderShareToStorySticker(true)
-      }
-      const encodedTrackId = encodeHashId(content.track.track_id)
-      const streamMp3Url = apiClient.makeUrl(`/tracks/${encodedTrackId}/stream`)
-      const storyVideoPath = path.join(
-        RNFS.TemporaryDirectoryPath,
-        `storyVideo-${uuid()}.mp4`
-      )
-      const audioStartOffsetConfig =
-        content.track.duration && content.track.duration >= 20 ? '-ss 10 ' : ''
-
-      let stickerUri: string | undefined // Have to capture the sticker image first because it doesn't work if you get the dominant colors first (mysterious).
-
-      try {
-        stickerUri = await captureStickerImage()
-      } catch (e) {
-        reportToSentry({
-          level: ErrorLevel.Error,
-          error: new Error(e),
-          name: 'Share to IG Story error - generate sticker step'
-        })
-        toast({ content: messages.shareToStoryError, type: 'error' })
-        return
-      }
-      if (!stickerUri) {
-        reportToSentry({
-          level: ErrorLevel.Error,
-          error: new Error('Sticker screenshot unsuccessful'),
-          name: 'Share to IG Story error - generate sticker step (sticker undefined)'
-        })
-        toast({ content: messages.shareToStoryError, type: 'error' })
-        return
-      }
-      let dominantColorsResult: Color[]
-      let dominantColorHex1: string
-      let dominantColorHex2: string
-      if (trackImageUri) {
-        try {
-          dominantColorsResult = await getDominantColors(trackImageUri)
-        } catch (e) {
-          reportToSentry({
-            level: ErrorLevel.Error,
-            error: e,
-            name: 'Share to IG Story error - calculate dominant colors step'
-          })
-          toast({ content: messages.shareToStoryError, type: 'error' })
-          return
-        }
-
-        ;[dominantColorHex1, dominantColorHex2] = Array.isArray(
-          dominantColorsResult
-        )
-          ? [
-              convertRGBToHex(dominantColorsResult[0]),
-              convertRGBToHex(dominantColorsResult[1])
-            ]
-          : DEFAULT_DOMINANT_COLORS
-      } else {
-        ;[dominantColorHex1, dominantColorHex2] = DEFAULT_DOMINANT_COLORS
-      }
-      let session: FFmpegSession
-      try {
-        ;[session] = await Promise.all([
-          FFmpegKit.execute(
-            `${audioStartOffsetConfig}-i ${streamMp3Url} -filter_complex "gradients=s=1080x1920:c0=${dominantColorHex1}:c1=${dominantColorHex2}:duration=10:speed=0.0225:rate=60[bg];[0:a]aformat=channel_layouts=mono,showwaves=mode=cline:n=1:s=1080x200:scale=cbrt:colors=#ffffff[fg];[bg][fg]overlay=format=auto:x=0:y=H-h-100" -pix_fmt yuv420p -vb 50M -t 10 ${storyVideoPath}`
-          )
-        ])
-      } catch (e) {
-        reportToSentry({
-          level: ErrorLevel.Error,
-          error: e,
-          name: 'Share to IG Story error'
-        })
-        toast({ content: messages.shareToStoryError, type: 'error' })
-        return
-      }
-
-      const returnCode = await session.getReturnCode()
-
-      if (!ReturnCode.isSuccess(returnCode)) {
-        const output = await session.getOutput()
-        reportToSentry({
-          level: ErrorLevel.Error,
-          error: new Error(output),
-          name: 'Share to IG Story error - generate video background step'
-        })
-        toast({ content: messages.shareToStoryError, type: 'error' })
-        return
-      }
-      const shareOptions = {
-        backgroundVideo: storyVideoPath,
-        stickerImage: stickerUri,
-        attributionURL: Config.AUDIUS_URL,
-        social: Share.Social.INSTAGRAM_STORIES,
-        appId: Config.INSTAGRAM_APP_ID
-      }
-      try {
-        await Share.shareSingle(shareOptions)
-      } catch (error) {
-        reportToSentry({
-          level: ErrorLevel.Error,
-          error,
-          name: 'Share to IG Story error - share to IG step'
-        })
-        toast({ content: messages.shareToStoryError, type: 'error' })
-      }
-    }
-  }, [
-    trackImageUri,
-    content,
-    captureStickerImage,
-    toast,
+  const {
+    handleShareToStoryStickerLoad,
+    handleShareToInstagramStory,
     shouldRenderShareToStorySticker
-  ])
+  } = useShareToStory({ content, viewShotRef })
 
   const handleCopyLink = useCallback(() => {
     if (!content) return
@@ -419,7 +253,8 @@ export const ShareDrawer = () => {
 
   return (
     <>
-      {shouldRenderShareToStorySticker ? (
+      {/* Redundant `content?.type === 'track'` needed to make TS happy */}
+      {content?.type === 'track' && shouldRenderShareToStorySticker ? (
         <ViewShot
           style={styles.viewShot}
           ref={viewShotRef}
@@ -427,10 +262,8 @@ export const ShareDrawer = () => {
         >
           <ShareToStorySticker
             onLoad={handleShareToStoryStickerLoad}
-            track={content.track}
-            // TODO(nkang): Not sure if this helps or hurts chances of successful image:
-            // user={account}
-            artist={content.artist}
+            track={content?.track}
+            artist={content?.artist}
           />
         </ViewShot>
       ) : null}
